requirements.txt
# CyberMind - GTK4 Mindmap Application
# System dependencies (install via dnf on Fedora 43):
#   sudo dnf install gtk4-devel libadwaita-devel python3-gobject cairo-devel
#
# Python dependencies:
PyGObject>=3.46.0
pycairo>=1.25.0
